+++ packages/field/src/components/Percent/util.ts
@@ -19,5 +19,5 @@ export function getColorByRealValue(realValue: number /** ,color: string */) {
 
 /** 获取到最后展示的数字 */
 export function getRealTextWithPrecision(realValue: number, precision: number = 2) {
-  return precision && precision > 0 ? realValue.toFixed(precision) : realValue;
+  return precision >= 0 ? realValue.toFixed(precision) : realValue;
 }