+++ packages/form/src/components/List/index.tsx
@@ -13,7 +13,6 @@ import { useEffect, useState } from 'react';
 import React, { useContext, useImperativeHandle, useMemo, useRef } from 'react';
 import './index.less';
 import { noteOnce } from 'rc-util/lib/warning';
-import type { StoreValue } from 'rc-field-form/es/interface';
 
 type IconConfig = {
   Icon?: React.FC<any>;
@@ -27,14 +26,6 @@ const FormListContext = React.createContext<
   | Record<string, any>
 >({});
 
-// type ChildrenFunction = (
-//   fields: FormListFieldData[],
-//   operation: FormListOperation,
-//   meta: {
-//     errors: React.ReactNode[];
-//   },
-// ) => React.ReactNode;
-
 type ChildrenItemFunction = (
   field: FormListFieldData,
   index: number,
@@ -43,11 +34,11 @@ type ChildrenItemFunction = (
 
 type FormListActionGuard = {
   beforeAddRow?: (
-    defaultValue?: StoreValue,
-    insertIndex?: number,
-    count?: number,
+    ...params: [...Parameters<FormListOperation['add']>, number]
   ) => boolean | Promise<boolean>;
-  beforeRemoveRow?: (index: number | number[], count?: number) => boolean | Promise<boolean>;
+  beforeRemoveRow?: (
+    ...params: [...Parameters<FormListOperation['remove']>, number]
+  ) => boolean | Promise<boolean>;
 };
 
 export type ProFormListProps = Omit<FormListProps, 'children'> & {
@@ -68,7 +59,7 @@ export type ProFormListProps = Omit<FormListProps, 'children'> & {
     defaultActionDom: ReactNode[],
     count: number,
   ) => ReactNode[];
-  children: ReactNode | ChildrenItemFunction;
+  children?: ReactNode | ChildrenItemFunction;
   itemContainerRender?: (
     doms: ReactNode,
     listMeta: {
@@ -86,6 +77,7 @@ export type ProFormListProps = Omit<FormListProps, 'children'> & {
   itemRender?: (
     doms: { listDom: ReactNode; action: ReactNode },
     listMeta: {
+      name: FormListProps['name'];
       field: FormListFieldData;
       fields: FormListFieldData[];
       index: number;
@@ -172,9 +164,19 @@ const ProFormListItem: React.FC<
     count,
     ...rest
   } = props;
+
   const listContext = useContext(FormListContext);
+
+  const unmountedRef = useRef(false);
+
   const [loadingRemove, setLoadingRemove] = useState(false);
   const [loadingCopy, setLoadingCopy] = useState(false);
+
+  useEffect(() => {
+    return () => {
+      unmountedRef.current = true;
+    };
+  }, []);
 
   const childrenArray = listToArray(children)
     .map((childrenItem) => {
@@ -183,10 +185,23 @@ const ProFormListItem: React.FC<
       }
       return childrenItem;
     })
-    .map((childrenItem) => {
+    .map((childrenItem, itemIndex) => {
       if (React.isValidElement(childrenItem)) {
+        const hasKey =
+          !!childrenItem.key ||
+          !!childrenItem?.props?.name ||
+          childrenItem?.type?.toString() === 'Symbol(react.fragment)';
+
+        noteOnce(
+          hasKey,
+          'ProFormList 的 children 不设置 key 可能导致更新不及时或者修改不生效的问题，请设置 key。',
+        );
+        noteOnce(
+          hasKey,
+          "ProFormList's children do not set the key may cause updates not to be timely or the modification does not take effect, please set the key.",
+        );
         return React.cloneElement(childrenItem, {
-          key: childrenItem.key || nanoid(),
+          key: childrenItem.key || childrenItem?.props?.name || itemIndex,
           ...childrenItem?.props,
         });
       }
@@ -241,13 +256,15 @@ const ProFormListItem: React.FC<
             onClick={async () => {
               setLoadingRemove(true);
               await action.remove(field.name);
-              setLoadingRemove(false);
+              if (!unmountedRef.current) {
+                setLoadingRemove(false);
+              }
             }}
           />
         </Spin>
       </Tooltip>
     );
-  }, [deleteIconProps, min, count, loadingRemove, prefixCls, action, field.name]);
+  }, [deleteIconProps, min, count, loadingRemove, prefixCls, setLoadingRemove, action, field.name]);
 
   const defaultActionDom: React.ReactNode[] = useMemo(
     () => [copyIcon, deleteIcon].filter(Boolean),
@@ -259,6 +276,7 @@ const ProFormListItem: React.FC<
   const dom = actions.length > 0 ? <div className={`${prefixCls}-action`}>{actions}</div> : null;
 
   const options = {
+    name: rest.name,
     field,
     index,
     record: formInstance?.getFieldValue?.(
@@ -288,6 +306,7 @@ const ProFormListItem: React.FC<
       {dom}
     </div>
   );
+
   return (
     <FormListContext.Provider
       key={field.name}
@@ -336,32 +355,17 @@ const ProFormListContainer: React.FC<ProFormListItemProps> = (props) => {
   const wrapperAction = useMemo(() => {
     const wrapAction = { ...action };
     const count = uuidFields.length;
-    Object.keys(action).forEach((key) => {
-      switch (key) {
-        case 'add':
-          wrapAction.add = async (defaultValue?: StoreValue, insertIndex?: number) => {
-            if (!actionGuard?.beforeAddRow) {
-              action.add(defaultValue, insertIndex);
-              return;
-            }
-            const needAdd = await actionGuard.beforeAddRow?.(defaultValue, insertIndex, count);
-            if (!needAdd) return;
-            action.add(defaultValue, insertIndex);
-          };
-          break;
-        case 'remove':
-          wrapAction.remove = async (idx: number | number[]) => {
-            if (!actionGuard?.beforeRemoveRow) {
-              action.remove(idx);
-              return;
-            }
-            const needRemove = await actionGuard.beforeRemoveRow?.(idx, count);
-            if (!needRemove) return;
-            action.remove(idx);
-          };
-          break;
-      }
-    });
+
+    if (actionGuard?.beforeAddRow) {
+      wrapAction.add = async (...rest) =>
+        (await actionGuard.beforeAddRow!(...rest, count)) && action.add(...rest);
+    }
+
+    if (actionGuard?.beforeRemoveRow) {
+      wrapAction.remove = async (...rest) =>
+        (await actionGuard.beforeRemoveRow!(...rest, count)) && action.remove(...rest);
+    }
+
     return wrapAction;
   }, [action, actionGuard, uuidFields]);
 
@@ -396,6 +400,7 @@ const ProFormListContainer: React.FC<ProFormListItemProps> = (props) => {
       style={{
         width: 'max-content',
         maxWidth: '100%',
+        minWidth: '100%',
       }}
     >
       {creatorButtonProps !== false && creatorButtonProps?.position === 'top' && creatorButton}
@@ -486,6 +491,7 @@ const ProFormList: React.FC<ProFormListProps> = ({
           {(fields, action, meta) => {
             // 将 action 暴露给外部
             actionRefs.current = action;
+
             return (
               <>
                 <ProFormListContainer