+++ packages/form/src/components/SchemaForm/layoutType/Embed.tsx
@@ -1,0 +1,3 @@
+const Embed: React.FC = ({ children }) => <>{children}</>;
+
+export default Embed;