+++ packages/form/src/components/Slider/index.tsx
@@ -1,16 +1,17 @@
 import React from 'react';
 import type { SliderSingleProps } from 'antd';
+import type { SliderRangeProps, SliderBaseProps } from 'antd/lib/slider';
 import { Slider } from 'antd';
 import type { ProFormFieldItemProps } from '../../interface';
 import { createField } from '../../BaseForm/createField';
 
-export type ProFormSliderProps = ProFormFieldItemProps<SliderSingleProps> & {
+export type ProFormSliderProps = ProFormFieldItemProps<SliderSingleProps | SliderRangeProps> & {
   range?: boolean;
-  min?: SliderSingleProps['min'];
-  max?: SliderSingleProps['max'];
-  step?: SliderSingleProps['step'];
-  marks?: SliderSingleProps['marks'];
-  vertical?: SliderSingleProps['vertical'];
+  min?: SliderBaseProps['min'];
+  max?: SliderBaseProps['max'];
+  step?: SliderBaseProps['step'];
+  marks?: SliderBaseProps['marks'];
+  vertical?: SliderBaseProps['vertical'];
 };
 /**
  * 文本选择组件