+++ packages/table/src/components/EditableTable/demos/basic.tsx
@@ -56,7 +56,7 @@ export default () => {
       tooltip: '只读，使用form.getFieldValue获取不到值',
       formItemProps: (form, { rowIndex }) => {
         return {
-          rules: rowIndex > 2 ? [{ required: true, message: '此项为必填项' }] : [],
+          rules: rowIndex > 1 ? [{ required: true, message: '此项为必填项' }] : [],
         };
       },
       // 第一行不允许编辑
@@ -147,6 +147,9 @@ export default () => {
         rowKey="id"
         headerTitle="可编辑表格"
         maxLength={5}
+        scroll={{
+          x: 960,
+        }}
         recordCreatorProps={
           position !== 'hidden'
             ? {
@@ -155,6 +158,7 @@ export default () => {
               }
             : false
         }
+        loading={false}
         toolBarRender={() => [
           <ProFormRadio.Group
             key="render"