+++ scripts/release.js
@@ -127,7 +127,7 @@ async function release() {
 
   process.env.NPM_CONFIG_OTP = otp;
 
-  pkgs.forEach((pkg, index) => {
+  const publishList = pkgs.map((pkg, index) => {
     const pkgPath = join(cwd, 'packages', pkg.replace('pro-', ''));
     const { name, version } = require(join(pkgPath, 'package.json'));
     const isNext = isNextVersion(version);
@@ -142,14 +142,16 @@ async function release() {
       console.log(
         `[${index + 1}/${pkgs.length}] Publish package ${name} ${isNext ? 'with next tag' : ''}`,
       );
-      const cliArgs = isNext ? ['publish', '--tag', 'next'] : ['publish'];
-      const { stdout } = execa.sync('npm', cliArgs, {
+      // 默认设置为 tag 检查通过之后在设置为 latest
+      const cliArgs = isNext ? ['publish', '--tag', 'next'] : ['publish', '--tag', 'beta'];
+      return execa('npm', cliArgs, {
         cwd: pkgPath,
       });
-      console.log(stdout);
     }
   });
-
+  console.log('发布中' + pkgs.join('/'));
+  await Promise.all(publishList);
+  console.log('发布成功！');
   await exec('npm', ['run', 'prettier']);
 
   logStep('done');